api/.vercel/cache/main/pyproject.toml
[project]
name = "app"
version = "0.1.0"
requires-python = ">=3.12"
classifiers = [
  "Private :: Do Not Upload",
]
dependencies = [
    "aiofiles>=23.0.0",
    "alembic>=1.17.0",
    "asyncpg>=0.29.0",
    "authlib>=1.6.5",
    "fastapi>=0.104.0",
    "httpx>=0.25.0",
    "itsdangerous>=2.1.0",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=10.0.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.5",
    "sqlalchemy>=2.0.0",
    "uvicorn[standard]>=0.24.0",
    "vercel-blob>=0.1.0",
    "werkzeug>=1.0.1",
]
